+++ connectome/__init__.py
@@ -1,2 +1,3 @@
 from .interface import *
+from .layers import *
 from .__version__ import __version__
