+++ connectome/containers/pipeline.py
@@ -1,7 +1,7 @@
 from .base import Container, EdgesBag
 
 
-class PipelineContainer(EdgesBag):
+class PipelineContainer(EdgesBag):  # pragma: no cover
     def __init__(self, head: EdgesBag, *tail: Container):
         self.containers = [head, *tail]
         for layer in tail:
@@ -9,13 +9,13 @@ class PipelineContainer(EdgesBag):
 
         state = head.freeze()
         super().__init__(state.inputs, state.outputs, state.edges, state.context, virtual_nodes=state.virtual_nodes,
-                         persistent_nodes=state.persistent_nodes)
+                         persistent_nodes=state.persistent_nodes, optional_nodes=None)
 
     def wrap(self, container: EdgesBag) -> EdgesBag:
         return PipelineContainer(container, *self.containers)
 
 
-class LazyPipelineContainer(Container):
+class LazyPipelineContainer(Container):  # pragma: no cover
     def __init__(self, *containers: Container):
         self.containers = containers
