+++ connectome/engine/__init__.py
@@ -1,2 +1,5 @@
-from .base import NodeHash
+from .base import Node, Edge, TreeNode, Request, Response, HashOutput, Command
+from .edges import ImpureEdge, CacheEdge, IdentityEdge, FunctionEdge, StaticEdge, StaticGraph
 from .executor import SyncExecutor, DefaultExecutor
+from .graph import Graph
+from .node_hash import TupleHash, NodeHashes, NodeHash, FilterHash
