+++ connectome/interface/__init__.py
@@ -1,6 +1,7 @@
-from .base import Chain, chained
 from .metaclasses import Source, Transform, Mixin
-from .blocks import CacheToRam, CacheToDisk, CacheColumns, Apply, Merge, Filter, GroupBy
+from .blocks import GroupBy
 from .decorators import *
 from .edges import *
 from .nodes import Input, Output, InverseInput, InverseOutput, Parameter
+# TODO: legacy
+from ..layers import Chain, chained, CacheToRam, CacheToDisk, CacheColumns, Apply, Merge, Filter  # noqa
