+++ connectome/layers/debug.py
@@ -1,0 +1,49 @@
+import hashlib
+from typing import Generator, Any
+
+from tarn.cache.storage import key_to_digest
+
+from ..containers.base import EdgesBag
+from ..engine.base import Node, Request, Response, Command
+
+from .cache import to_seq
+from ..engine.edges import StaticGraph, StaticHash
+from ..engine.node_hash import NodeHashes, NodeHash
+from ..utils import StringsLike
+
+from .base import CallableLayer
+
+
+class HashDigest(CallableLayer):
+    def __init__(self, names: StringsLike, algorithm):
+        if isinstance(algorithm, str):
+            algorithm = getattr(hashlib, algorithm)
+
+        names = to_seq(names)
+        inputs, outputs, edges = [], [], []
+        for name in names:
+            inp, out = Node(name), Node(name)
+            inputs.append(inp)
+            outputs.append(out)
+            edges.append(HashDigestEdge(algorithm).bind(inp, out))
+
+        super().__init__(EdgesBag(
+            inputs, outputs, edges,
+            context=None, virtual_nodes=None, persistent_nodes=None, optional_nodes=None,
+        ), ())
+
+
+class HashDigestEdge(StaticGraph, StaticHash):
+    def __init__(self, algorithm):
+        super().__init__(arity=1)
+        self.algorithm = algorithm
+
+    def _make_hash(self, inputs: NodeHashes) -> NodeHash:
+        return inputs[0]
+
+    def evaluate(self) -> Generator[Request, Response, Any]:
+        value = yield Command.ParentValue, 0
+        output = yield Command.CurrentHash,
+
+        pickled, digest = key_to_digest(self.algorithm, output.value)
+        return value, output.value, digest, pickled
