+++ tests/test_interface/test_chain.py
@@ -1,11 +1,9 @@
 from pathlib import Path
 
 import pytest
-from connectome import Source, Transform, Chain, CacheToRam, meta
-from connectome.exceptions import DependencyError, FieldError
-from connectome.interface.base import LazyChain
-from connectome.interface.blocks import HashDigest, CacheColumns, Merge
 from tarn.config import StorageConfig, init_storage
+from connectome import Source, Transform, Chain, CacheToRam, meta, LazyChain, HashDigest, CacheColumns, Merge
+from connectome.exceptions import DependencyError, FieldError
 
 
 class DS(Source):
@@ -35,7 +33,7 @@ def test_chain():
 def test_nested(block_maker):
     one = block_maker.first_ds(first_constant=2, ids_arg=15)
     two = block_maker.crop()
-    hash_layer = HashDigest('image')
+    hash_layer = HashDigest('image', 'blake2b')
 
     base, *variants = [
         Chain(one, two, hash_layer),
