+++ tests/test_interface/test_filter.py
@@ -17,7 +17,7 @@ def test_filter(block_maker):
     ids = pipeline.ids
     assert ids == ('4', '14')
 
-    hash_layer = HashDigest(['image', 'lungs', 'spacing'])
+    hash_layer = HashDigest(['image', 'lungs', 'spacing'], 'blake2b')
     hashed = Chain(block, hash_layer)
     pipeline = Chain(
         block, Filter(lambda image: image.endswith('4')), hash_layer,
